Move EXTLINUX to pyanaconda.bootloader.extlinux
Move the EXTLINUX class to a new file.
